+++ test/tempo/explain_test.exs
@@ -12,7 +12,7 @@ defmodule Tempo.Explain.Test do
   # output.
 
   setup_all do
-    Calendar.put_time_zone_database(Tzdata.TimeZoneDatabase)
+    Calendar.put_time_zone_database(Tz.TimeZoneDatabase)
     :ok
   end