+++ test/tempo/ical_test.exs
@@ -38,7 +38,7 @@ defmodule Tempo.ICal.Test do
                             end)
 
   setup_all do
-    Calendar.put_time_zone_database(Tzdata.TimeZoneDatabase)
+    Calendar.put_time_zone_database(Tz.TimeZoneDatabase)
     :ok
   end